requirements.txt
arrow==1.2.2
asgiref==3.5.2
binaryornot==0.4.4
boto3==1.24.26
botocore==1.27.26
certifi==2022.6.15
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
cookiecutter==2.1.1
Django==4.0.6
django-cors-headers==3.13.0
django-storages==1.12.3
djangorestframework==3.13.1
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.13.2
hyperframe==6.0.1
idna==3.3
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath==1.0.1
MarkupSafe==2.1.1
Pillow==9.2.0
priority==2.0.0
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.1
PyYAML==6.0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tzdata==2022.1
urllib3==1.26.10
wsproto==1.1.0
